fix: remove login rsc, move to client, reorganise checks
