+++ package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "build": "next build",
     "dev": "next dev",
-    "start": "cp -r .next/static .next/standalone/.next/ && cp .env .next/standalone && cp -r ./public .next/standalone/public && node .next/standalone/server.js",
+    "start": "cp -r .next/static .next/standalone/.next/ && cp -r ./public .next/standalone/public && node .next/standalone/server.js",
     "format:check": "prettier --ignore-path .gitignore --list-different '**/**/*.{ts,json,md}'",
     "format:fix": "npm run format:check -- --write",
     "lint": "eslint . --ext .ts,.tsx",
